backend/requirements.txt
fastapi>=0.115.8
uvicorn>=0.30.6
python-dotenv>=1.0.1
httpx>=0.27.2
python-multipart>=0.0.12
pydantic>=2.10.6
pydantic-settings>=2.6.1
supabase>=2.11.0
python-jose>=3.3.0
cachetools>=5.5.0
casparser>=0.7.4
pdfplumber>=0.11.4
google-generativeai>=0.8.4
yfinance>=0.2.54
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.22.0
